+++ client/src/components/UI/Button.tsx
@@ -1,4 +1,4 @@
-import styled from 'styled-components';
+import { styled } from 'styled-components';
 
 const Button = ({ children }) => {
   return <BtnComponent type="button">{children}</BtnComponent>;
