+++ client/src/components/UI/Input.tsx
@@ -1,4 +1,4 @@
-import styled from 'styled-components';
+import { styled } from 'styled-components';
 
 const Input = () => {
   return <InputComponent type="text"></InputComponent>;
