+++ client/src/components/UI/Radio.tsx
@@ -1,4 +1,4 @@
-import styled from 'styled-components';
+import { styled } from 'styled-components';
 
 const Radio = () => {
   return (
