+++ client/src/components/UI/Tag.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import styled from 'styled-components';
+import { styled } from 'styled-components';
 
 const Tag = () => {
   const [isClicked, setIsClicked] = useState(true);
