+++ client/src/components/layout/Header.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 import { Link } from 'react-router-dom';
-import styled from 'styled-components';
+import { styled } from 'styled-components';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faLocationDot, faBars, faX } from '@fortawesome/free-solid-svg-icons';
-import HeaderNavLogin from '../buttons/HeaderNavLogin.tsx';
+import HeaderNavLogin from './HeaderNavLogin.tsx';
 
 const Header = () => {
   const [isClicked, setIsClicked] = useState(true);
