+++ client/src/constant/constant.ts
@@ -1,0 +1,13 @@
+export const GENDER_TAGS = [
+  { id: 1, text: '# 여자만' },
+  { id: 2, text: '# 남자만' },
+  { id: 3, text: '# 남녀노소' },
+];
+
+export const FOOD_TAGS = [
+  { id: 1, text: '# 한식' },
+  { id: 2, text: '# 중식' },
+  { id: 3, text: '# 양식' },
+  { id: 4, text: '# 일식' },
+  { id: 5, text: '# 기타' },
+];
