+++ client/src/fetch/postSignUp.tsx
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-const SignUpPost = async (email: string, password: string, copyPassword: string) => {
+const postSignUp = async (email: string, password: string, copyPassword: string) => {
   try {
     const response = await axios.post('/users/signup', { email, password, copyPassword });
     return response.data;
@@ -8,4 +8,4 @@ const SignUpPost = async (email: string, password: string, copyPassword: string)
     throw new Error('이미 가입 되어있는 이메일 입니다.'); // You can customize the error message as per your requirements
   }
 };
-export default SignUpPost;
+export default postSignUp;
